+++ CHANGELOG.md
@@ -9,10 +9,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - (Keep your changes here until you have a release version)
 
 ## [0.1.25] 2020-12-07
+### Added
+- Red asterisk to indicate required fields
+
 ### Changed
 - Notes look in UI to look more like post-it notes
 - Notes section of Large Request View now contains status update notes with colors associated to that statuses
 - Regular users can now add notes to their own Requests
+- Reject Review buttons disabled when no comment is provided
 
 ### Removed
 - Ability for users to edit/delete notes